Resi

+++ JRTParcel/database/migrations/2024_06_09_151931_create_barang_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('resi_id');
             $table->string('tipe_komoditas');
-            $table->integer('berat');
+            $table->float('berat');
             $table->float('lebar');
             $table->float('panjang');
             $table->float('tinggi');
